Disable ONNX export for older torch versions
